+++ core/src/main/kotlin/kt/firmata/core/protocol/fsm/I2CMessageEvent.kt
@@ -1,3 +1,3 @@
 package kt.firmata.core.protocol.fsm
 
-data class I2CMessageEvent(val address: Int, val register: Int, val message: ByteArray) : Event
+data class I2CMessageEvent(val address: Int, val register: Int, val message: IntArray) : Event
